refactor: Update file structure and component imports in View files
