+++ frontend/src/components/View/Common/CanvasContainer.js
