+++ frontend/src/components/View/Common/ResetButton.js
