+++ frontend/src/components/View/Common/ViewTitle.js
