+++ frontend/src/components/View/Common/WelcomeMessage.js
